requirements.txt
Django>=4.2
psycopg2
requests
youtube-transcript-api
python-dotenv
beautifulsoup4
gunicorn